PoC readonlyMounts with flags
This is Proof of Concept of how we could get readonly mounts
to have "parent" flags. I personally don't like how much
operations it is required to this solution. I thought that I
will be able to use "data" as coma-separated list of parameters
but it doesn't work so now I would need to translate names into
flags. From command line one could just run:
findmnt -T /proc/
and it would return the mount options but here we are iteraing
through.